# Improve performance of LifecycleObjectCache