--- a/src/StructureMap/Building/PushPopWrapper.cs
+++ b/src/StructureMap/Building/PushPopWrapper.cs
@@ -2,14 +2,13 @@
 using System.Linq.Expressions;
 using System.Reflection;
 using StructureMap.Pipeline;
-using StructureMap.TypeRules;
 
 namespace StructureMap.Building
 {
     public static class PushPopWrapper
     {
-        public static readonly MethodInfo PushMethod = typeof (IBuildSession).GetMethod("Push");
-        public static readonly MethodInfo PopMethod = typeof (IBuildSession).GetMethod("Pop");
+        public static readonly MethodInfo PushMethod = typeof(IBuildSession).GetMethod(nameof(IBuildSession.Push));
+        public static readonly MethodInfo PopMethod = typeof(IBuildSession).GetMethod(nameof(IBuildSession.Pop));
 
         public static Expression WrapFunc(Type returnType, Instance instance, Expression inner)
         {